requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
decorator==5.2.1
duckduckgo_search==8.0.2
Flask==3.1.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.2.6
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
primp==0.15.0
proglog==0.1.12
pypiwin32==223
pytesseract==0.3.13
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
tqdm==4.67.1
Werkzeug==3.1.3
